+++ tests/tests_unit/tests_common/test_utils.py
@@ -1,0 +1,19 @@
+import pytest
+from datetime import datetime
+from th2_data_services_lwdp.utils import _check_milliseconds
+
+
+def test_milliseconds():
+    assert _check_milliseconds(datetime.now().replace(microsecond=0)) is None
+
+
+def test_milliseconds_invalid_type():
+    with pytest.raises(Exception) as err:
+        _check_milliseconds(datetime.now().timestamp())
+    assert "Provided timestamp should be `datetime` object" in str(err)
+
+
+def test_milliseconds_invalid_time():
+    with pytest.raises(Exception) as err:
+        _check_milliseconds(datetime.now())
+    assert "Provided datetime shouldn't contain microseconds" in str(err)